requirements.txt
Flask==1.1.2
google-cloud-datastore==1.15.3
requests==2.24.0
requests-oauthlib==1.3.0